my-k8s-api/requirements-dev.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
kubernetes==32.0.1
requests==2.32.3
aiofiles==24.1.0
pydantic==2.11.4
Jinja2==3.1.6
python-multipart==0.0.20
# 開發環境最小依賴，不包含 CUDA/GPU 套件